requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
bcrypt==4.3.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
build==1.2.2.post1
    # via pip-tools
click==8.2.1
    # via
    #   flask
    #   pip-tools
flask==3.1.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==25.0
    # via
    #   build
    #   gunicorn
pip-tools==7.4.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
werkzeug==3.1.3
    # via flask
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
